Do not remove files that have been edited when doing an "update".  Instead, mark the files as CONFLICT.
FossilOrigin-Name: fe6ee8a431b69e5695801bd703fb6a38ce30dc40

+++ src/update.c
@@ -146,11 +146,15 @@ void update_cmd(void){
       printf("UPDATE %s\n", zName);
       vfile_to_disk(0, idt, 0);
     }else if( idt==0 && idv>0 ){
-      char *zFullPath;
-      printf("REMOVE %s\n", zName);
-      zFullPath = mprintf("%s/%s", g.zLocalRoot, zName);
-      unlink(zFullPath);
-      free(zFullPath);
+      if( chnged ){
+        printf("CONFLICT %s\n", zName);
+      }else{
+        char *zFullPath;
+        printf("REMOVE %s\n", zName);
+        zFullPath = mprintf("%s/%s", g.zLocalRoot, zName);
+        unlink(zFullPath);
+        free(zFullPath);
+      }
     }else if( idt>0 && idv>0 && ridt!=ridv && chnged ){
       /* Merge the changes in the current tree into the target version */
       Blob e, r, t, v;
